main.py
import cv2
from stitching import Stitcher
import numpy as np

def stitch_microscope_images(images):
    stitcher = Stitcher()
    panorama = stitcher.stitch(images)
    return panorama

def count_cell_colonies(panorama):
    # Initiate ORB detector
    orb = cv2.ORB_create()

    # Find the keypoints and descriptors with ORB
    kp, des = orb.detectAndCompute(panorama, None)

    # ***THE FEATURE DESCRIPTION HERE IS INCOMPLETE YET. NEEDS REVISIONS.***
    # Placeholder: Perform feature matching and analyze extracted features to identify cell colonies
    # Replace this with your actual implementation to count cell colonies
    num_colonies = len(kp) // 50  # Placeholder counting logic

    return num_colonies

def main():
    # Assuming microscope images are passed as input arguments
    images = [cv2.imread(image) for image in input("Enter image filepaths separated by space: ").split()]

    # Stitch microscope images together to form a panorama
    panorama = stitch_microscope_images(images)

    # Count cell colonies in the panorama using ORB feature extraction
    num_colonies = count_cell_colonies(panorama)

    print("Number of cell colonies:", num_colonies)

if __name__ == "__main__":
    main()