requirements.txt
openstitching==1.0.0
numpy>=1.19.0
opencv-python>=4.5.3